Chuangmi Plug V1 and V3 merged.
